requirements.txt
Flask
Pillow
numpy
tensorflow-cpu
keras
Markup
jinja2
jsonpickle
gunicorn
gtts
opencv-python-headless
yaml
